LPS-80663 Make it a inner class, no logic change
