+++ portal-impl/src/com/liferay/portal/poller/PollerServlet.java
@@ -16,10 +16,15 @@ package com.liferay.portal.poller;
 
 import com.liferay.petra.string.StringPool;
 import com.liferay.portal.kernel.exception.NoSuchLayoutException;
+import com.liferay.portal.kernel.json.JSONArray;
+import com.liferay.portal.kernel.json.JSONFactoryUtil;
 import com.liferay.portal.kernel.json.JSONObject;
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
+import com.liferay.portal.kernel.notifications.ChannelException;
 import com.liferay.portal.kernel.notifications.ChannelHubManagerUtil;
+import com.liferay.portal.kernel.notifications.ChannelListener;
+import com.liferay.portal.kernel.notifications.NotificationEvent;
 import com.liferay.portal.kernel.poller.PollerHeader;
 import com.liferay.portal.kernel.servlet.ServletResponseUtil;
 import com.liferay.portal.kernel.util.ContentTypes;
@@ -28,6 +33,10 @@ import com.liferay.portal.kernel.util.PortalUtil;
 import com.liferay.portal.util.PropsValues;
 
 import java.io.IOException;
+
+import java.util.List;
+import java.util.concurrent.CountDownLatch;
+import java.util.concurrent.TimeUnit;
 
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServlet;
@@ -119,5 +128,48 @@ public class PollerServlet extends HttpServlet {
 	}
 
 	private static final Log _log = LogFactoryUtil.getLog(PollerServlet.class);
+
+	private static class SynchronousPollerChannelListener
+		implements ChannelListener {
+
+		@Override
+		public void channelListenerRemoved(long channelId) {
+			_countDownLatch.countDown();
+		}
+
+		public String getNotificationEvents(
+				long companyId, long userId,
+				JSONObject pollerResponseHeaderJSONObject, long timeout)
+			throws ChannelException {
+
+			try {
+				_countDownLatch.await(timeout, TimeUnit.MILLISECONDS);
+			}
+			catch (InterruptedException ie) {
+			}
+
+			List<NotificationEvent> notificationEvents =
+				ChannelHubManagerUtil.fetchNotificationEvents(
+					companyId, userId, true);
+
+			JSONArray jsonArray = JSONFactoryUtil.createJSONArray();
+
+			jsonArray.put(pollerResponseHeaderJSONObject);
+
+			for (NotificationEvent notificationEvent : notificationEvents) {
+				jsonArray.put(notificationEvent.toJSONObject());
+			}
+
+			return jsonArray.toString();
+		}
+
+		@Override
+		public void notificationEventsAvailable(long channelId) {
+			_countDownLatch.countDown();
+		}
+
+		private final CountDownLatch _countDownLatch = new CountDownLatch(1);
+
+	}
 
 }
